Cleanup imports, move af to top level dir
